Clean up the make frame file by abstracting the process of creating a frame to the handle Frame types New file frame controls, which will soon render the controls for each frame, delete, switch axis, save etc
