+++ src/js/ui/NewCharting/ChartGlobalControls.js
@@ -16,7 +16,8 @@ export default function ChartGlobalControls(props) {
                 <Button onClick={() => props.make({ type: ChartingType.HISTOGRAM })}>Histogram</Button>
                 <Button onClick={() => props.make({ type: ChartingType.SCATTERPLOT })}>Scatterplot</Button>
                 <Button onClick={() => props.make({ type: ChartingType.LINE })}>COVID-19</Button>
-                <Button onClick={() => props.make({ type:  ChartingType.PIEGRAPH })}>Pie Graph</Button>
+                <Button onClick={() => props.make({ type:  ChartingType.BOXPLOT })}>Box Plot</Button>
+                <Button onClick={() => props.make({ type:  ChartingType.RADARCHART })}>Radar Chart</Button>
             </ButtonGroup>
 
             {/* Close button */}
