+++ src/js/ui/NewCharting/ChartTemplates/RadarChart.js
@@ -1,0 +1,7 @@
+import React, { useEffect, useState } from 'react';
+
+
+export default function RadarChart(props) {
+
+
+}
